extract java part into parseJavaTemplates.ts
